+++ firmware/boron-ins-fsm/src/imDoorSensor.cpp
@@ -19,6 +19,7 @@ bool doorMessageReceivedFlag = false;
 unsigned long doorHeartbeatReceived = 0;
 unsigned long doorLastMessage = 0;
 unsigned long timeWhenDoorClosed = 0;
+unsigned long consecutiveOpenDoorHeartbeatCount = 0;
 
 //**********setup()******************
 
@@ -74,15 +75,23 @@ doorData checkIM() {
         // Checks if the 0th bit (counting from 0) of doorStatus is 1
         doorTamperedFlag = (currentDoorData.doorStatus & 0b0001) != 0;
 
-        // Reset timer on receiving a door close message or transition from open to closed + heartbeat
         if ((currentDoorData.doorStatus & 0b0010) == 0) {
+            // Reset timer on receiving a door close message or transition from open to closed + heartbeat
             if ((currentDoorData.doorStatus & 0b1000) == 0 || (previousDoorData.doorStatus & 0b0010) != 0) {
                 timeWhenDoorClosed = millis();
             }
+
+            // reset consecutive door open counter since the door is closed
+            consecutiveOpenDoorHeartbeatCount = 0;
         }
 
         // After a certain thereshold, the next door message received will trigger a boron heartbeat
         if (millis() - doorLastMessage >= MSG_TRIGGER_SM_HEARTBEAT_THRESHOLD) {
+            // if the door is open upon sending this heartbeat, increment count
+            if (isDoorOpen(currentDoorData.doorStatus)) {
+                consecutiveOpenDoorHeartbeatCount++;
+            }
+
             doorMessageReceivedFlag = true;
         }