+++ firmware/boron-ins-fsm/src/stateMachine.cpp
@@ -437,6 +437,9 @@ void getHeartbeat() {
         // Add "isINSZero" field to the JSON message
         writer.name("isINSZero").value(isINSZero && lastHeartbeatPublish > 0);
 
+        // sends consecutive heartbeat count to server side to be handled
+        writer.name("consecutiveOpenDoorHeartbeatCount").value(consecutiveOpenDoorHeartbeatCount);
+
         if (didMissQueue.size() > SM_HEARTBEAT_DID_MISS_QUEUE_SIZE) {
             // if oldest value did miss; subtract from the current amount
             if (didMissQueue.front()) {